change id type from int to bigint

+++ repository/src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 create table category(
-	id int not null AUTO_INCREMENT,
+	id bigint not null AUTO_INCREMENT,
 	name varchar(40) not null,
-	parent_id int,
+	parent_id bigint,
 	activate smallint,
 	created_time timestamp,
 	created_by varchar(20),
@@ -12,7 +12,7 @@ create table category(
 );
 
 create table subject(
-	id int not null auto_increment,
+	id bigint not null auto_increment,
 	name varchar(40) not null,
 	showOnIndex smallint,
 	activate smallint,
@@ -20,7 +20,7 @@ create table subject(
 );
 
 create table site_user(
-	id int not null auto_increment,
+	id bigint not null auto_increment,
 	name varchar(20),
 	password char(64),
 	enabled smallint,
@@ -31,19 +31,19 @@ create table site_user(
 );
 
 create TABLE provider(
-	id int NOT NULL AUTO_INCREMENT,
+	id bigint NOT NULL AUTO_INCREMENT,
 	PRIMARY KEY (id),
 	FOREIGN KEY (id) REFERENCES site_user(id)
 );
 
 create table common_user(
-	id int not null auto_increment,
+	id bigint not null auto_increment,
 	primary key (id),
 	foreign key (id) references site_user(id)
 );
 
 create table brand(
-	id int not null auto_increment,
+	id bigint not null auto_increment,
 	name varchar(50),
 	created_time timestamp,
 	acitvate smallint,
@@ -51,18 +51,18 @@ create table brand(
 );
 
 create table product(
-	id int not null auto_increment,
+	id bigint not null auto_increment,
 	name varchar(20) not null,
 	description varchar(300),
 	cover_img varchar(40),
 	private float,
-	first_category_id int,
-	second_category_id int,
-	third_category_id int,
+	first_category_id bigint,
+	second_category_id bigint,
+	third_category_id bigint,
 	price float,
 	url varchar(50),
-	brand_id int,
-	created_by int,
+	brand_id bigint,
+	created_by bigint,
 	primary key(id),
 	foreign key (first_category_id) references category(id),
 	foreign key (second_category_id) references category(id),
@@ -72,32 +72,32 @@ create table product(
 );
 
 create table consumable(
-	id int not null,
+	id bigint not null,
 	primary key(id),
 	foreign key(id) references product(id)
 );
 
 create table instrument(
-	id int not null,
+	id bigint not null,
 	primary key(id),
 	foreign key(id) references product(id)
 );
 
 create table reagents(
-	id int not null,
+	id bigint not null,
 	primary key(id),
 	foreign key(id) references product(id)
 );
 
 create table service(
-	id int not null,
+	id bigint not null,
 	primary key(id),
 	foreign key(id) references product(id)
 );
 
 create table product_subject(
-	product_id int,
-	subject_id int,
+	product_id bigint,
+	subject_id bigint,
 	added_on timestamp,
 	sort_order int,
 	primary key(product_id,subject_id),
@@ -107,8 +107,8 @@ create table product_subject(
 
 
 create table prefer_products(
-	user_id int not null,
-	product_id int not null,
+	user_id bigint not null,
+	product_id bigint not null,
 	ADDED_ON timestamp,
 	primary key(user_id,product_id),
 	foreign key(user_id) references common_user(id),
